Extract Kinds & update imports
