+++ src/UI/Routing/Kinds.ts
@@ -1,0 +1,8 @@
+export type Kinds =
+  | "Catalog"
+  | "Inventory"
+  | "CreateInstance"
+  | "History"
+  | "Diagnose"
+  | "Events"
+  | "Resources";
